+++ libs/ngx.command/src/command.model.ts
@@ -1,16 +1,32 @@
-import { Observable } from "rxjs";
+import type { Observable } from "rxjs";
+import type { Signal } from "@angular/core";
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type ExecuteFn = (...args: any[]) => unknown;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type ExecuteAsyncFn = (...args: any[]) => Observable<unknown> | Promise<unknown>;
+export type CanExecute = (() => boolean) | Signal<boolean> | Observable<boolean>;
 
 export interface ICommand {
-	/** Determines whether the command is currently executing, as a snapshot value. */
+	/** Determines whether the command is currently executing, as a snapshot value.
+	 * @deprecated Use {@link $isExecuting} signal instead.
+	*/
 	readonly isExecuting: boolean;
-	/** Determines whether the command is currently executing, as an observable. */
-	readonly isExecuting$: Observable<boolean>;
-	/** Determines whether the command can execute or not, as a snapshot value. */
-	readonly canExecute: boolean;
-	/** Determines whether the command can execute or not, as an observable. */
-	readonly canExecute$: Observable<boolean>;
 
-	/** Determines whether to auto destroy when having 0 subscribers (defaults to `true`). */
+	/** Determines whether the command is currently executing, as a signal. */
+	readonly $isExecuting: Signal<boolean>;
+
+	/** Determines whether the command can execute or not, as a snapshot value.
+	 * @deprecated Use {@link $canExecute} signal instead.
+	*/
+	readonly canExecute: boolean;
+
+	/** Determines whether the command can execute or not, as a signal. */
+	readonly $canExecute: Signal<boolean>;
+
+	/** Determines whether to auto destroy when having 0 subscribers (defaults to `true`).
+	 * @deprecated Use using command/commandAsync is handled automatically.
+	 */
 	autoDestroy: boolean;
 
 	/** Execute function to invoke. */
@@ -19,10 +35,15 @@ export interface ICommand {
 	/** Disposes all resources held by subscriptions. */
 	destroy(): void;
 
-	/** Subscribe listener, in order to handle auto disposing. */
+	/** Subscribe listener, in order to handle auto disposing.
+	 * @deprecated Use using command/commandAsync is handled automatically.
+	 */
 	subscribe(): void;
 
-	/** Unsubscribe listener, in order to handle auto disposing. */
+	/**
+	 * Unsubscribe listener, in order to handle auto disposing.
+	 * @deprecated Use using command/commandAsync is handled automatically.
+	*/
 	unsubscribe(): void;
 }
 
@@ -31,8 +52,9 @@ export interface CommandCreator {
 	// execute: (...args: TParams extends unknown[] ? TParams : [TParams]) => Observable<unknown> | Promise<unknown> | void;
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	execute: (...args: any[]) => Observable<unknown> | Promise<unknown> | void;
+	// CanExecute | ((...args: any[]) => CanExecute);
 	// eslint-disable-next-line @typescript-eslint/no-unsafe-function-type
-	canExecute?: Observable<boolean> | Function;
+	canExecute?: CanExecute | Function;
 	params?: unknown | unknown[];
 	isAsync?: boolean;
 	host: unknown;