+++ libs/ngx.command/src/command.ts
@@ -1,15 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import {
-	Observable, combineLatest, Subscription, Subject, BehaviorSubject, of, EMPTY,
-	tap, map, filter, switchMap, catchError, finalize, take,
+	Observable, Subscription, Subject, of, EMPTY,
+	tap, filter, switchMap, catchError, finalize, take,
 } from "rxjs";
-import type { ICommand } from "./command.model";
-import { assertInInjectionContext, computed, DestroyRef, inject, Injector, isSignal, type Signal } from "@angular/core";
-import { toObservable } from "@angular/core/rxjs-interop";
-
-export type ExecuteFn = (...args: any[]) => unknown;
-export type ExecuteAsyncFn = (...args: any[]) => Observable<unknown> | Promise<unknown>;
-export type CanExecute = (() => boolean) | Signal<boolean> | Observable<boolean>;
+import { toSignal } from "@angular/core/rxjs-interop";
+import type { CanExecute, ExecuteAsyncFn, ExecuteFn, ICommand } from "./command.model";
+import { assertInInjectionContext, computed, DestroyRef, inject, Injector, isSignal, signal, type Signal } from "@angular/core";
 
 export interface CommandCreateOptions {
 	isAsync: boolean,
@@ -43,7 +39,7 @@ export function command(
 	const injector = opts?.injector ?? inject(Injector);
 	const isAsync = opts?.isAsync ?? false;
 	const destroyRef = injector.get(DestroyRef);
-	const cmd = new Command(execute, canExecute$, isAsync);
+	const cmd = new Command(execute, canExecute$, isAsync, injector);
 	cmd.autoDestroy = false;
 
 	destroyRef.onDestroy(() => {
@@ -55,37 +51,22 @@ export function command(
 
 /**
  * Command object used to encapsulate information which is needed to perform an action.
- * @deprecated Use {@link command} or {@link commandAsync} instead for creating instances.
+ *
  */
 export class Command implements ICommand {
 
-	/** Determines whether the command is currently executing, as a snapshot value. */
-	get isExecuting(): boolean {
-		return this._isExecuting;
-	}
+	get isExecuting(): boolean { return this.$isExecuting(); }
 
-	/** Determines whether the command can execute or not, as a snapshot value. */
-	get canExecute(): boolean {
-		return this._canExecute;
-	}
+	get canExecute(): boolean { return this.$canExecute(); }
 
-	/** Determines whether the command is currently executing, as an observable. */
-	get isExecuting$(): Observable<boolean> {
-		return this._isExecuting$.asObservable();
-	}
+	readonly $isExecuting = signal(false);
+	readonly $canExecute = computed(() => !this.$isExecuting() && this._$canExecute());
 
-	/** Determines whether to auto destroy when having 0 subscribers. */
+	private readonly _$canExecute: Signal<boolean>;
+
 	autoDestroy = true;
 
-	/** Determines whether the command can execute or not, as an observable. */
-	readonly canExecute$: Observable<boolean>;
-
-	private _isExecuting$ = new BehaviorSubject<boolean>(false);
-	private _isExecuting = false;
-	private _canExecute = true;
 	private executionPipe$ = new Subject<unknown[] | undefined>();
-	private isExecuting$$ = Subscription.EMPTY;
-	private canExecute$$ = Subscription.EMPTY;
 	private executionPipe$$ = Subscription.EMPTY;
 	private subscribersCount = 0;
 
@@ -95,6 +76,7 @@ export class Command implements ICommand {
 	 * @param execute Execute function to invoke - use `isAsync: true` when `Observable<any>`.
 	 * @param canExecute Observable which determines whether it can execute or not.
 	 * @param isAsync Indicates that the execute function is async e.g. Observable.
+	 * @deprecated Use {@link command} or {@link commandAsync} instead for creating instances.
 	 */
 	constructor(
 		execute: ExecuteFn,
@@ -106,29 +88,11 @@ export class Command implements ICommand {
 			const canExecute = typeof canExecute$ === "function"
 				? computed(canExecute$)
 				: canExecute$;
-			this.canExecute$ = combineLatest([
-				this._isExecuting$,
-				isSignal(canExecute) ? toObservable(canExecute, { injector }) : canExecute
-			]).pipe(
-				map(([isExecuting, canExecuteResult]) => {
-					// console.log("[command::combineLatest$] update!", { isExecuting, canExecuteResult });
-					this._isExecuting = isExecuting;
-					this._canExecute = !isExecuting && !!canExecuteResult;
-					return this._canExecute;
-				}),
-			);
-			this.canExecute$$ = this.canExecute$.subscribe();
+			this._$canExecute = isSignal(canExecute)
+				? canExecute
+				: toSignal(canExecute, { initialValue: false, injector });
 		} else {
-			this.canExecute$ = this._isExecuting$.pipe(
-				map(x => {
-					const canExecute = !x;
-					this._canExecute = canExecute;
-					return canExecute;
-				})
-			);
-			this.isExecuting$$ = this._isExecuting$
-				.pipe(tap(x => this._isExecuting = x))
-				.subscribe();
+			this._$canExecute = signal(true);
 		}
 		this.executionPipe$$ = this.buildExecutionPipe(execute, isAsync).subscribe();
 	}
@@ -143,8 +107,6 @@ export class Command implements ICommand {
 	destroy(): void {
 		// console.warn("[command::destroy]");
 		this.executionPipe$$.unsubscribe();
-		this.canExecute$$.unsubscribe();
-		this.isExecuting$$.unsubscribe();
 	}
 
 	subscribe(): void {
@@ -162,10 +124,10 @@ export class Command implements ICommand {
 	private buildExecutionPipe(execute: (...args: unknown[]) => any, isAsync?: boolean): Observable<unknown> {
 		let pipe$ = this.executionPipe$.pipe(
 			// tap(x => console.warn(">>>> executionPipe", this._canExecute)),
-			filter(() => this._canExecute),
+			filter(() => this.$canExecute()),
 			tap(() => {
 				// console.log("[command::executionPipe$] do#1 - set execute", { args: x });
-				this._isExecuting$.next(true);
+				this.$isExecuting.set(true);
 			})
 		);
 
@@ -189,7 +151,7 @@ export class Command implements ICommand {
 				execFn,
 				finalize(() => {
 					// console.log("[command::executionPipe$]  finalize inner#1 - set idle");
-					this._isExecuting$.next(false);
+					this.$isExecuting.set(false);
 				}),
 				take(1),
 				catchError(error => {
@@ -197,12 +159,11 @@ export class Command implements ICommand {
 					return EMPTY;
 				}),
 			)),
-			tap(
-				() => {
-					// console.log("[command::executionPipe$] tap#2 - set idle");
-					this._isExecuting$.next(false);
-				},
-			)
+			tap(() => {
+				// console.log("[command::executionPipe$] tap#2 - set idle");
+				// this._isExecuting$.next(false);
+				this.$isExecuting.set(false);
+			}),
 		);
 		return pipe$;
 	}
@@ -216,6 +177,9 @@ export class Command implements ICommand {
  */
 export class CommandAsync extends Command {
 
+	/**
+	 * @deprecated Use {@link commandAsync} instead to create an instance.
+	 */
 	constructor(
 		execute: ExecuteAsyncFn,
 		canExecute$?: CanExecute,