+++ libs/ngx.command/src/index.ts
@@ -1,8 +1,8 @@
 export { SsvCommandModule } from "./command.module";
 export { provideSsvCommandOptions, COMMAND_OPTIONS, type CommandOptions } from "./command.options";
-export { command, commandAsync, Command, CommandAsync, type CanExecute, type CommandCreateOptions } from "./command";
-export { CommandDirective } from "./command.directive";
-export { CommandRefDirective } from "./command-ref.directive";
+export { command, commandAsync, Command, CommandAsync, type CommandCreateOptions } from "./command";
+export { SsvCommand } from "./command.directive";
+export { SsvCommandRef } from "./command-ref.directive";
 export { type CanExecuteFormOptions, isCommand, isCommandCreator, canExecuteFromNgForm, canExecuteFromSignals } from "./command.util";
-export type { CommandCreator, ICommand } from "./command.model";
+export type { CommandCreator, ICommand, CanExecute, ExecuteAsyncFn, ExecuteFn } from "./command.model";
 export { VERSION } from "./version";